Review the logic around ROWCOUNT in qgen.
We should be able to use qgen in order to produce a different query each
time, following the distribution properties set by TPC, and which depend on
the Scale Factor, if I read correctly the README.

Fix the logic so that the query generated is actually usable with
PostgreSQL, so that the following command now works:

    DSS_QUERY=queries ./qgen -s 1 -a 1 | psql -a -d tpch

It seems like what's intended here.

In passing, I coudn't resist doing some Emacs re-indent of the function, for
my own sanity, Sorry if that makes the patch too heavy.