Implement direct loading mode for PostgreSQL.
The load_stub.c trick from the source code almost works for that. Almost,
because its API insists that operations are done on a line-per-line basis
without too much though given to batching operations.

So this patch adds some parameters in the loader function prototypes and
pass them from driver.c to the two implementations of them, that are print.c
and load_stub.c. Maybe this file should be renames pg_direct.c or something.

With this patch it's now possible to issue a direct load command that will
connect to a PostgreSQL and use COPY to stream generated data directly into
our instance:

  $ time ./dbgen -n postgresql:///tpch -D -s 30000 -C 30000 -S 1 -v
  TPC-H Population Generator (Version 2.13.0)
  Copyright Transaction Processing Performance Council 1994 - 2010
          Starting to load stage 1 of 30000 for suppliers table.../
  Preloading text ... 100%
  done.
          Starting to load stage 1 of 30000 for customers table...done.
          Starting to load stage 1 of 30000 for orders/lineitem tables...done.
          Starting to load stage 1 of 30000 for part/partsupplier tables...done.
  Generating data for nation tabledone.
  Generating data for region tabledone.

  real    0m47.222s
  user    0m31.940s
  sys     0m7.204s

Now, this patch might be too invasive for citus/test-automation repository,
depending on what we want it to contain in terms ot TPC-H driver. To be
discussed, this “fork” of TPC-H with the PostgreSQL driver could well live
in another repo.