Add support for DELETE in direct mode.
This implements the 2.7 Old Sales Refresh Function (RF2) part of the specs
in Direct Load with PostgreSQL, issuing the DELETE statements directly to
the database.

The 2.6 New Sales Refresh Function (RF1) query was already supported thanks
to reusing the internal tdefs structure, which is switched to the direct
load support as a whole when using -D. The pr_range() function isn't part of
the tdef struct, though, so required specific care.