add hammerdb tpcc scripts
build.tcl and run.tcl are added, which are used to build the hammerdb
tables and run the actual benchmark. The reason we have to have two
different scripts rather than one is that we run some other commands
after building the tables, such as distributing functions in citus etc.