add scripts to run tpcc in an automated way
In order to be able to run the test in an automated way, screen is used
to create a detached session, github token and username are cached in
the driver node so that when it needs to access a private repository, it
can use the tokens.

The create-and-run.sh script will create the cluster and the driver node
in the same subnet, so that the network latency will be lower.