add a script to run the analytical queries
A script to run the extra 22 queries is added. The script can use
multiple threads to send the analytical queries. For each thread, it
will start from a random index, and it will continue sequentially from
that index. The seed for random is given, which should be the same seed
when testing across different platforms to be fair. Some of the queries
take much longer compared to others.

Each query run time will be logged so that we can see how much a certain
query takes. The script catches the kill signal and logs the results.