store public ip in storage account
Each node in the cluster will store its public ip in the storage account
that is used in the cluster. Then all of the public ips from the storage
account will be added as trusted so that each node in the cluster can
connect to another node without requiring ssh-forwarding.