refactor hammerdb scripts to be more generic
The hammerdb version will be given as an argument in create-run script.
Then it will be written to HAMMERDB_VERSION file so that other hammerdb
scripts can get the version from that file instead of hardcoded version.

Instead of personal repository, citusdata is used for hammerdb fork.
This has some small changes related to distributing tables at the
beginning, which increases speed.

Max connections is increased to 600, because in hyperscale we not have
600 max connections in workers. The only difference is that now the
coordinator will also have 600 max connections even though in hyperscale
coordinator has 300 max connections. This shouldn't matter that much
though.