Disable deadlock prevention in hammerdb config
With the latest hammerdb changes (4.0), there are multi-shard updates.
Citus will use ShareUpdateExclusiveLock lock mode instead of
RowExclusiveLock if deadlock prevention is enabled. The locks are local
therefore there is not point in this prevention unless the updates are
always done from the same node. Ideally we should disable it only on the
workers but we currently don't have that structure in test automation
hence for hammerdb we will disable it in the cluster.