Add a job to delete resource group for hammerdb automatically
A job to delete a resource group added. If a branch starts with
delete_resource_group/ it will delete the resource group by extracting
it from the branch name.

Hammerdb job will push a branch to test automation to trigger this job
so that we don't need to manually delete the resource group after
hammerdb benchmarks are done.

For example, if the resource group name is citusbot_resource_group_name,
hammerdb will push the following branch to test automation after
completion: delete_resource_group/<id>_citusbot_resource_group_name and
the job will extract the resource group name to delete it.