use centos in our vms
We were using redhat in our vms but it turns out to have some problems.
Currently centos seems to be running smoothly. So switching to centos
makes sense, and it doesn't require many changes. We can go back to
redhat or even make it configurable easily.

In another PR the versions will be increased to the latest 7.X.