log start time of each query and log checkpoints
It is useful to log checkpoints so that we know when they are run and
how they affect the runtimes. In order to understand how they affect the
runtime we should also know when each query started. So that is also
added.
In terms of performance there shouldn't be any difference as there won't
be lots of checkpoint logs.