retry more times in case of timeout error
We have an extra security rule in azure vms and we delete that rule
before sshing, however sometimes the rule comes back too quickly and we
get a timeout while trying to ssh. Normally we would retry 4 times but
sometimes the issue is persistent for a bit more, so increasing it makes
sense.