Fix init.sh to retry connecting to redhat repositories (#279)
This PR fixes the problem about `create-cluser.sh` failing to
automatically provision clusters for testing.

Because azure is deploying resources in parallel, and at the time
custom-script extension runs init.sh, redhat repositories could not be
resolved. This is most probably because network resources are not fully
deployed during that time.
 
To test the theory, I placed a while loop in the init.sh which just
tries to install hostname package until it succeeds, or sleeps 10
seconds if it fails, and retries that some 50 times.

With this patch, `create-cluster.sh` can now succeed.

---------

Co-authored-by: Alper Kocatas <alperkocatas@microsoft.com>